feat: carry the catering credential to the phone
The Dart half of the credential sync: a `KuchniaCredentialService` that owns
the value locally, a `sync_merge_kuchnia.dart` adapter mirroring the Python
one, and a `_syncKuchniaCredential` tick alongside the banks and budget. The
phone can now receive the panel password, which is the prerequisite for it
fetching anything itself.

One source of truth, per the answered question: the value lives in the synced
document store and nowhere else. Deliberately *not* mirrored into
`flutter_secure_storage` -- two homes for one value is how a stale password
survives a change made on the other device, and the credential has to be
readable synchronously at merge time anyway.

Gated by a second shared fixture, `tests/fixtures/kuchnia_credential.json`,
asserted by both suites. Two cases carry the weight:

- an **unparsable** edit time must fall back to the same clock on both sides
  (Python catches `ValueError` from `fromisoformat`, Dart's `tryParse` returns
  null). Verified they agree on `wall_time_ms == 0`; mutating that expectation
  fails both suites.
- a **blank half** contributes nothing, so a half-filled settings form cannot
  push an empty-string password that wins LWW against a real one.

`KuchniaCredentialService.init()` is called in the WorkManager isolate as well
as `main()`. A fresh isolate has its own static state and `_syncKuchniaCredential`
no-ops when the singleton is missing, so without it a background tick would
neither publish this device's credential nor receive a peer's -- and on the
phone the background tick is most of the syncing there is. Same class of trap
as the `initSyncDeviceId()` one already documented there.

`.toDouble()` came off the bank record: the fields are already `double`, so it
was a no-op the analyzer flags. The field *types* are what keep the encoding
byte-identical to Python's, and `kuchnia_parity_test.dart` asserts that.

654 Dart tests, 1040 Python tests, 100% Python branch coverage.

+++ app/lib/models/kuchnia_dish.dart
@@ -58,17 +58,20 @@ class KuchniaDish {
   /// ranks foods by how often they were *eaten*, and a delivered dish has not
   /// been eaten yet.
   ///
-  /// Every macro is forced to `double` before it reaches `jsonEncode`.
+  /// Every macro is declared `double` on this class rather than `num`, which
+  /// is what keeps the encoded record byte-identical to Python's.
   /// `jsonEncode(435)` emits `435` where Python emits `435.0`, and a
   /// byte-different record for the same dish makes every refresh look like a
-  /// change -- which republishes the whole curated bank to every peer.
+  /// change -- which republishes the whole curated bank to every peer. A
+  /// `.toDouble()` here would be a no-op the analyzer flags; the field types
+  /// are the guarantee, and `kuchnia_parity_test.dart` asserts the encoding.
   Map<String, dynamic> toBankRecord() => {
     'desc': name,
-    'kcal': kcal.toDouble(),
-    'protein_g': proteinG.toDouble(),
-    'carbs_g': carbsG.toDouble(),
-    'fat_g': fatG.toDouble(),
-    'grams': grams.toDouble(),
+    'kcal': kcal,
+    'protein_g': proteinG,
+    'carbs_g': carbsG,
+    'fat_g': fatG,
+    'grams': grams,
     'count': 0,
   };
